Fix review findings on the unresponsive-friend throttle (PR #345)
- Guard the per-friend outbox read in sendLocation()'s filter so one
  friend's store error doesn't abort the batch send for every friend.
- Extract the duplicated in-memory MailboxClient test fake used by three
  test files into a single shared MemoryMailboxClient.
